improve circular_dependency test
